fixed issues with animation
I will be updating the animation code with a more flexible solution later on. I also intend to implement a greedy algorithm for the color assignments for extra clarity and a few extra algorithms along with a drop down menu. Lastly, I will erase some of the unused or redundant variables.